bgpd:Fix for AS_PATH size calculation for long AS Paths
If you have an AS_PATH with more entries than
what can be written into a single AS_SEGMENT_MAX
it needs to be broken up.  The code that noticed
that the AS_PATH needs to be broken up was not
correctly calculating the size of the resulting
message.  This patch addresses this issue.

Reported by: Andrew Kerr <andrew.kerr@openmarket.com>