The problem is that bgpd dumps the Large Communities into the MRT file with an attribute type code of BGP_ATTR_COMMUNITIES instead of BGP_ATTR_LARGE_COMMUNITIES
As a result, if the RIB entry also have regular communities, the
resulting MRT file has two community attributes (which is a protocol
violation, to have the same attribute more than once).

In the case of bgpdump, this crashes it with an assert. For some other
parsers, they only display the second attribute (losing the first set of
communities) and parse the large communities as if they were regular
communities.